refactor: 아키텍처 구조 재설계
- 이 과정에서 도커파일 추가
