+++ .gitignore
@@ -3,3 +3,6 @@
 
 ### macOS ###
 .DS_Store
+
+### VScode ###
+.vscode
