+++ Dockerfile
@@ -1,18 +1,22 @@
-# 베이스 이미지
+# Base image
 FROM python:3.11-slim
 
-# 작업 디렉토리 설정
+# Set working directory
 WORKDIR /app
 
-# 종속성 설치
+# Install dependencies
 COPY requirements.txt .
 RUN pip install --no-cache-dir -r requirements.txt
 
-# 애플리케이션 파일 복사
+# Copy application files
 COPY . .
 
-# 포트 노출
+# Expose port
 EXPOSE 8000
 
-# FastAPI 실행 명령
-CMD ["uvicorn", "api:app", "--host", "0.0.0.0", "--port", "8000"]
+# Environment variables
+# Note: ELEVENLABS_API_KEY should be provided at runtime
+ENV PYTHONUNBUFFERED=1
+
+# Run FastAPI application
+CMD ["uvicorn", "api.api:app", "--host", "0.0.0.0", "--port", "8000"]
