+++ app/templates/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Text-to-Speech Web App</title>
-    <link rel="stylesheet" href="/static/css/style.css">
+    <link rel="stylesheet" href="/app/staticyle.css">
 </head>
 <body>
     <div class="container">
@@ -37,6 +37,6 @@
         </div>
     </div>
     
-    <script src="/static/js/script.js"></script>
+    <script src="/app/staticipt.js"></script>
 </body>
 </html>
